refactor: extract TimerSubscription interface
The implementation of the TimerSubscription is specific to delayed
timers, where the scheduling in ActorTimerQueue determines the actual
timestamp to use.

I want to add a new TimerSubscription implementation that isn't based on
delays but simply uses a static timestamp as deadline.

Initially, I looked into replacing the delay code of the
TimerSubscription, but this isn't possible as it also allows recurring
timers using the same delay for each trigger.

Therefore, I chose to extract an interface, allowing me to more
re-implement the class differently without affecting existing behavior.

(cherry picked from commit 5d7066c73e1b47d583ba5cac741b95fc74d4a523)
