+++ scheduler/src/main/java/io/camunda/zeebe/scheduler/ActorThread.java
@@ -146,7 +146,7 @@ public class ActorThread extends Thread implements Consumer<Runnable> {
   }
 
   /** Must be called from this thread, schedules a job to be run later. */
-  public void scheduleTimer(final TimerSubscription timer) {
+  public void scheduleTimer(final DelayedTimerSubscription timer) {
     timerJobQueue.schedule(timer, clock);
   }
