+++ scheduler/src/main/java/io/camunda/zeebe/scheduler/TimerSubscription.java
@@ -9,121 +9,35 @@ package io.camunda.zeebe.scheduler;
 
 import java.util.concurrent.TimeUnit;
 
-public final class TimerSubscription implements ActorSubscription, ScheduledTimer, Runnable {
-  private final ActorJob job;
-  private final ActorTask task;
-  private final TimeUnit timeUnit;
-  private final long deadline;
-  private final boolean isRecurring;
-  private volatile boolean isDone = false;
-  private volatile boolean isCanceled = false;
-  private long timerId = -1L;
-  private ActorThread thread;
-  private long timerExpiredAt;
-
-  public TimerSubscription(
-      final ActorJob job, final long deadline, final TimeUnit timeUnit, final boolean isRecurring) {
-    this.job = job;
-    task = job.getTask();
-    this.timeUnit = timeUnit;
-    this.deadline = deadline;
-    this.isRecurring = isRecurring;
-  }
+public interface TimerSubscription extends ActorSubscription, ScheduledTimer, Runnable {
 
   @Override
-  public boolean poll() {
-    return isDone;
-  }
+  boolean poll();
 
   @Override
-  public ActorJob getJob() {
-    return job;
-  }
+  ActorJob getJob();
 
   @Override
-  public boolean isRecurring() {
-    return isRecurring;
-  }
+  boolean isRecurring();
 
   @Override
-  public void onJobCompleted() {
-
-    if (isRecurring && !isCanceled) {
-      isDone = false;
-      submit();
-    }
-  }
+  void onJobCompleted();
 
   @Override
-  public void cancel() {
-    if (!isCanceled && (!isDone || isRecurring)) {
-      task.onSubscriptionCancelled(this);
-      isCanceled = true;
-      final ActorThread current = ActorThread.current();
+  void cancel();
 
-      if (current != thread) {
-        thread.submittedCallbacks.add(this);
-      } else {
-        run();
-      }
-    }
-  }
+  long getTimerId();
 
-  public long getTimerId() {
-    return timerId;
-  }
+  void setTimerId(long timerId);
 
-  public void setTimerId(final long timerId) {
-    this.timerId = timerId;
-  }
+  void submit();
 
-  public void submit() {
-    thread = ActorThread.current();
-    thread.scheduleTimer(this);
-  }
+  long getDeadline();
 
-  public long getDeadline() {
-    return deadline;
-  }
-
-  public TimeUnit getTimeUnit() {
-    return timeUnit;
-  }
-
-  public void onTimerExpired(final TimeUnit timeUnit, final long now) {
-    if (!isCanceled) {
-      isDone = true;
-      timerExpiredAt = timeUnit.toNanos(now);
-      task.tryWakeup();
-    }
-  }
+  void onTimerExpired(TimeUnit timeUnit, long now);
 
   @Override
-  public void run() {
-    thread.removeTimer(this);
-  }
+  void run();
 
-  @Override
-  public String toString() {
-    return "TimerSubscription{"
-        + "timerId="
-        + timerId
-        + ", deadline="
-        + deadline
-        + ", timeUnit="
-        + timeUnit
-        + ", isRecurring="
-        + isRecurring
-        + ", isDone="
-        + isDone
-        + ", isCanceled="
-        + isCanceled
-        + ", thread="
-        + thread
-        + '}';
-  }
-
-  public long getTimerExpiredAt() {
-    return timerExpiredAt;
-  }
+  long getTimerExpiredAt();
 }
